Revert "Pages shield (#3918)"
This reverts commit 53f94a9f824a60292a300b8eac9a43ea22818dd2.
