+++ lib/plausible_web/controllers/site_controller.ex
@@ -257,7 +257,7 @@ defmodule PlausibleWeb.SiteController do
   end
 
   def settings_shields(conn, %{"shield" => shield})
-      when shield in ["ip_addresses", "countries", "pages"] do
+      when shield in ["ip_addresses", "countries"] do
     site = conn.assigns.site
 
     conn
