+++ lib/plausible_web/templates/site/settings_shields.html.heex
@@ -14,11 +14,4 @@
         "domain" => @site.domain
       }
     ) %>
-  <% "pages" -> %>
-    <%= live_render(@conn, PlausibleWeb.Live.Shields.Pages,
-      session: %{
-        "site_id" => @site.id,
-        "domain" => @site.domain
-      }
-    ) %>
 <% end %>
