+++ lib/plausible_web/views/layout_view.ex
@@ -65,15 +65,10 @@ defmodule PlausibleWeb.LayoutView do
       %{
         key: "Shields",
         icon: :shield_exclamation,
-        value:
-          [
-            %{key: "IP Addresses", value: "shields/ip_addresses"},
-            %{key: "Countries", value: "shields/countries"},
-            if FunWithFlags.enabled?(:shield_pages) do
-              %{key: "Pages", value: "shields/pages"}
-            end
-          ]
-          |> Enum.reject(&is_nil/1)
+        value: [
+          %{key: "IP Addresses", value: "shields/ip_addresses"},
+          %{key: "Countries", value: "shields/countries"}
+        ]
       },
       %{key: "Email Reports", value: "email-reports", icon: :envelope},
       if conn.assigns[:current_user_role] == :owner do
